Changes before error encountered
Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/9d72b34a-34b5-472b-996a-7f6216fe7cb6
