+++ src/format/src/printer/print.ts
@@ -1,25 +1,24 @@
 /**
  * Central print dispatcher for the GML formatter workspace.
  *
- * ARCHITECTURE NOTE: This file has grown organically and now houses both high-level
- * print coordination logic and low-level node-handling utilities. It should be refactored
- * into multiple focused modules:
+ * Split history
+ * -------------
+ * This file has been progressively refactored.  Functions that were once
+ * inlined here now live in focused sub-modules:
  *
- * - A top-level coordinator that delegates to domain-specific sub-printers
- * - Separate files for each AST node category (expressions, statements, declarations, etc.)
- * - General AST utilities (node inspection, property access) should move to Core
- * - Comment handling should be extracted to a dedicated comment-printer module
+ * - doc-comment-output.ts     – doc-comment block printing for function/variable nodes
+ * - expression-print-utils.ts – parenthesis-flattening, ternary printing, and shared
+ *                               expression utility helpers (printSimpleDeclaration,
+ *                               printEmptyParens, printEmptyBlock, etc.)
  *
- * Until this refactoring occurs, contributors should avoid adding new utility functions
- * here; instead, place domain-specific helpers in appropriately-scoped files under
- * src/format/src/printer/ or src/core/src/ast/ and import them as needed.
+ * Contributors should continue placing new domain-specific helpers in the appropriate
+ * sub-module rather than growing this file further.
  */
 
-import { Core, type MutableDocCommentLines } from "@gmloop/core";
+import { Core } from "@gmloop/core";
 import { util } from "prettier";
 
 import { printComment, printDanglingComments, printDanglingCommentsAsGroup } from "../comments/comment-printer.js";
-import { buildPrintableDocCommentLines } from "../comments/description-doc.js";
 import {
     LogicalOperatorsStyle,
     normalizeLogicalOperatorsStyle,
@@ -27,9 +26,20 @@ import {
     resolveObjectWrapOption,
     TRAILING_COMMA
 } from "../options/index.js";
-import { MULTIPLICATIVE_BINARY_OPERATORS, NUMBER_TYPE, OBJECT_TYPE, STRING_TYPE, UNDEFINED_TYPE } from "./constants.js";
+import { DOC_COMMENT_OUTPUT_FLAG, NUMBER_TYPE, OBJECT_TYPE, STRING_TYPE, UNDEFINED_TYPE } from "./constants.js";
+import { markDocCommentsAsPrinted, printNodeDocComments } from "./doc-comment-output.js";
 import { getEnumNameAlignmentPadding, prepareEnumMembersForPrinting } from "./enum-alignment.js";
-import { safeGetParentNode, safeGetPathName, safeGetPathValue } from "./path-utils.js";
+import {
+    docHasTrailingComment,
+    printEmptyBlock,
+    printEmptyParens,
+    printSimpleDeclaration,
+    printTernaryExpressionNode,
+    printWithoutExtraParens,
+    shouldBreakVariableInitializerOnAssignmentLine,
+    shouldOmitSyntheticParens
+} from "./expression-print-utils.js";
+import { safeGetParentNode } from "./path-utils.js";
 import {
     breakParent,
     concat,
@@ -55,7 +65,6 @@ import {
 import { buildClauseGroup, printSingleClauseStatement } from "./single-clause-statement.js";
 import {
     getOriginalTextFromOptions,
-    hasBlankLineBetweenLastCommentAndClosingBrace,
     macroTextHasExplicitTrailingBlankLine,
     resolveNodeIndexRangeWithSource,
     resolvePrinterSourceMetadata,
@@ -64,22 +73,15 @@ import {
 } from "./source-text.js";
 import { shouldAddNewlinesAroundStatement, shouldSuppressEmptyLineBetween } from "./statement-spacing-policy.js";
 import {
-    expressionIsStringLike,
-    hasLineBreak,
     isCallbackArgument,
     isComplexArgumentNode,
-    isInlineEmptyBlockComment,
     isInLValueChain,
-    isNumericComputationNode,
-    isSimpleCallArgument,
-    isSyntheticParenFlatteningEnabled
+    isSimpleCallArgument
 } from "./type-guards.js";
 import { joinDeclaratorPartsWithCommas } from "./variable-declarator-layout.js";
 
 const forcedStructArgumentBreaks = new WeakMap();
 const MIN_VARIABLE_DECLARATIONS_BEFORE_LOOP_PADDING = 4;
-
-const DOC_COMMENT_OUTPUT_FLAG = "_gmlHasDocCommentOutput";
 
 function applyLogicalOperatorsStyle(operator, style) {
     const coreStyle = style === LogicalOperatorsStyle.KEYWORDS ? "keyword" : "symbol";
@@ -206,226 +208,6 @@ function tryPrintFunctionNode(node, path, options, print) {
     const body = printFunctionBody(node, path, options, print);
 
     return concat([docComments, signature, " ", body]);
-}
-
-function printNodeDocComments(node, path, options) {
-    const sourceMetadata = resolvePrinterSourceMetadata(options);
-    const { originalText } = sourceMetadata;
-    const { startIndex: nodeStartIndex } = resolveNodeIndexRangeWithSource(node, sourceMetadata);
-
-    const docCommentDocs: MutableDocCommentLines = Array.isArray(node.docComments)
-        ? Core.toMutableArray(node.docComments as string[], { clone: true })
-        : [];
-    const plainLeadingLines: string[] = Array.isArray(node.plainLeadingLines) ? node.plainLeadingLines : [];
-
-    // The formatter trusts the AST's `docComments` as authoritative. Legacy doc
-    // comment formats (e.g. `// @function`) are normalised by the lint rule
-    // `gml/normalize-doc-comments` before formatting, so no source-text fallback
-    // is needed here. Core's `normalizeFunctionDocCommentAttachments` helper
-    // pre-attaches recognised `@function`-tag comments to the correct function
-    // node, removing the need for any formatter-side source-text scan.
-    // (target-state.md §2.2, §3.2, §3.5)
-
-    sortDocCommentsBySourceOrder(docCommentDocs);
-
-    const docCommentEntriesForMetadata = [...docCommentDocs];
-    const printableDocComments = buildPrintableDocCommentLines(docCommentDocs, originalText);
-    const printableDocCommentBlock = joinDocCommentsPreservingSourceSpacing(
-        printableDocComments,
-        docCommentEntriesForMetadata,
-        originalText
-    );
-
-    const parts: any[] = [];
-    const shouldEmitPlainLeadingLines = plainLeadingLines.length > 0;
-
-    if (shouldEmitPlainLeadingLines) {
-        parts.push(join(hardline, plainLeadingLines), hardline);
-        if (docCommentDocs.length === 0) {
-            parts.push(hardline);
-        }
-    }
-
-    if (docCommentDocs.length > 0) {
-        node[DOC_COMMENT_OUTPUT_FLAG] = true;
-        const suppressLeadingBlank = (docCommentDocs as any)?._suppressLeadingBlank === true;
-
-        const needsLeadingBlankLine = node?._gmlNeedsLeadingBlankLine === true;
-
-        const hasLeadingNonDocComment =
-            !Core.isNonEmptyArray(node.docComments) &&
-            docCommentDocs.length === 0 &&
-            originalText !== null &&
-            typeof nodeStartIndex === NUMBER_TYPE &&
-            Core.hasCommentImmediatelyBefore(originalText, nodeStartIndex);
-
-        const hasExistingBlankLine =
-            originalText !== null &&
-            typeof nodeStartIndex === NUMBER_TYPE &&
-            util.isPreviousLineEmpty(originalText, nodeStartIndex);
-        const isTopOfFileDocBlock =
-            originalText !== null &&
-            typeof nodeStartIndex === NUMBER_TYPE &&
-            originalText.slice(0, nodeStartIndex).trim().length === 0;
-
-        const shouldEmitConfiguredLeadingBlankLine =
-            !suppressLeadingBlank &&
-            ((!isTopOfFileDocBlock && needsLeadingBlankLine) || (hasLeadingNonDocComment && !hasExistingBlankLine));
-
-        if (shouldEmitConfiguredLeadingBlankLine) {
-            parts.push(hardline);
-        }
-
-        parts.push(printableDocCommentBlock, hardline);
-    } else {
-        if (Object.hasOwn(node, DOC_COMMENT_OUTPUT_FLAG)) {
-            delete node[DOC_COMMENT_OUTPUT_FLAG];
-        }
-    }
-
-    markDocCommentsAsPrinted(node, path);
-
-    return concat(parts);
-}
-
-function joinDocCommentsPreservingSourceSpacing(
-    printableDocComments: ReadonlyArray<unknown>,
-    docCommentDocs: MutableDocCommentLines,
-    originalText: string | null
-) {
-    if (!Core.isNonEmptyArray(printableDocComments)) {
-        return "";
-    }
-
-    if (originalText === null || printableDocComments.length !== docCommentDocs.length) {
-        return join(hardline, [...printableDocComments] as any[]);
-    }
-
-    const parts: any[] = [];
-    for (let index = 0; index < printableDocComments.length; index += 1) {
-        parts.push(printableDocComments[index]);
-
-        if (index >= printableDocComments.length - 1) {
-            continue;
-        }
-
-        const currentEntry = docCommentDocs[index];
-        const nextEntry = docCommentDocs[index + 1];
-        if (hasBlankLineBetweenDocCommentEntries(currentEntry, nextEntry, originalText)) {
-            parts.push(hardline, hardline);
-        } else {
-            parts.push(hardline);
-        }
-    }
-
-    return concat(parts);
-}
-
-function hasBlankLineBetweenDocCommentEntries(leftEntry: unknown, rightEntry: unknown, originalText: string): boolean {
-    const leftEndIndex = resolveDocCommentEndIndex(leftEntry);
-    const rightStartIndex = resolveDocCommentStartIndex(rightEntry);
-    if (leftEndIndex === null || rightStartIndex === null || rightStartIndex <= leftEndIndex) {
-        return false;
-    }
-
-    const slice = originalText.slice(leftEndIndex + 1, rightStartIndex);
-    if (slice.length === 0) {
-        return false;
-    }
-
-    return /\r?\n[ \t]*\r?\n/u.test(slice);
-}
-
-function resolveDocCommentStartIndex(commentEntry: unknown): number | null {
-    if (!Core.isObjectLike(commentEntry)) {
-        return null;
-    }
-
-    const startValue = (commentEntry as { start?: unknown }).start;
-    if (typeof startValue === NUMBER_TYPE) {
-        return startValue as number;
-    }
-
-    if (Core.isObjectLike(startValue)) {
-        const startIndex = (startValue as { index?: unknown }).index;
-        if (typeof startIndex === NUMBER_TYPE) {
-            return startIndex as number;
-        }
-    }
-
-    return null;
-}
-
-function resolveDocCommentEndIndex(commentEntry: unknown): number | null {
-    if (!Core.isObjectLike(commentEntry)) {
-        return null;
-    }
-
-    const endValue = (commentEntry as { end?: unknown }).end;
-    if (typeof endValue === NUMBER_TYPE) {
-        return endValue as number;
-    }
-
-    if (Core.isObjectLike(endValue)) {
-        const endIndex = (endValue as { index?: unknown }).index;
-        if (typeof endIndex === NUMBER_TYPE) {
-            return endIndex as number;
-        }
-    }
-
-    return null;
-}
-
-function sortDocCommentsBySourceOrder(docCommentDocs: MutableDocCommentLines): void {
-    if (!Array.isArray(docCommentDocs) || docCommentDocs.length <= 1) {
-        return;
-    }
-
-    const indexedEntries = docCommentDocs.map((entry, index) => ({
-        entry,
-        index,
-        startIndex: resolveDocCommentStartIndex(entry)
-    }));
-
-    const hasSourcePositions = indexedEntries.some((entry) => typeof entry.startIndex === NUMBER_TYPE);
-    if (!hasSourcePositions) {
-        return;
-    }
-
-    indexedEntries.sort((left, right) => {
-        const leftStart = typeof left.startIndex === NUMBER_TYPE ? left.startIndex : Number.NEGATIVE_INFINITY;
-        const rightStart = typeof right.startIndex === NUMBER_TYPE ? right.startIndex : Number.NEGATIVE_INFINITY;
-        if (leftStart !== rightStart) {
-            return leftStart - rightStart;
-        }
-        return left.index - right.index;
-    });
-
-    for (const [index, indexedEntry] of indexedEntries.entries()) {
-        docCommentDocs[index] = indexedEntry.entry;
-    }
-}
-
-function markDocCommentsAsPrinted(node, path) {
-    if (node.docComments) {
-        node.docComments.forEach((comment: any) => {
-            if (comment && typeof comment === "object") {
-                comment.printed = true;
-            }
-        });
-    } else {
-        const parentNode = safeGetParentNode(path);
-        if (parentNode && parentNode.type === Core.VARIABLE_DECLARATOR) {
-            const grandParentNode = safeGetParentNode(path, 1);
-            if (grandParentNode && grandParentNode.type === Core.VARIABLE_DECLARATION && grandParentNode.docComments) {
-                grandParentNode.docComments.forEach((comment: any) => {
-                    if (comment && typeof comment === "object") {
-                        comment.printed = true;
-                    }
-                });
-            }
-        }
-    }
 }
 
 function printFunctionSignature(node, path, options, print) {
@@ -2681,599 +2463,3 @@ function buildIfAlternateDoc(path, options, print, node) {
     return printInBlock(path, options, print, "alternate");
 }
 
-function docHasTrailingComment(doc) {
-    if (!Core.isNonEmptyArray(doc)) {
-        return false;
-    }
-
-    const lastItem = doc.at(-1);
-    if (!Core.isNonEmptyArray(lastItem)) {
-        return false;
-    }
-
-    const commentArr = lastItem[0];
-    if (!Core.isNonEmptyArray(commentArr)) {
-        return false;
-    }
-
-    return commentArr.some((item) => {
-        return typeof item === STRING_TYPE && (item.startsWith("//") || item.startsWith("/*"));
-    });
-}
-
-function printWithoutExtraParens(path, print, ...keys) {
-    return path.call((childPath) => unwrapParenthesizedExpression(childPath, print), ...keys);
-}
-
-function getBinaryOperatorInfo(operator) {
-    if (operator === undefined) {
-        return;
-    }
-    return Core.BINARY_OPERATORS[operator];
-}
-
-function shouldOmitSyntheticParens(path, _options) {
-    void _options;
-    const node = safeGetPathValue(path);
-    if (!node || node.type !== "ParenthesizedExpression") {
-        return false;
-    }
-
-    // Focus on synthetic parentheses (those inserted by the parser or formatter for
-    // precedence disambiguation) rather than explicit parentheses written by the
-    // user. Removing user-written parentheses could alter intended grouping or
-    // emphasis, while synthetic ones exist solely to clarify operator precedence
-    // and can be safely omitted when the context makes precedence unambiguous.
-    const isSynthetic = node.synthetic === true;
-
-    const parent = safeGetParentNode(path);
-    if (!parent) {
-        return false;
-    }
-
-    const parentKey = safeGetPathName(path);
-    const expression = node.expression;
-
-    if (shouldStripStandaloneAdditiveParentheses(parent, parentKey, expression)) {
-        return true;
-    }
-
-    if (parent.type === "TernaryExpression") {
-        return shouldFlattenTernaryTest(parentKey, expression);
-    }
-
-    // Always strip redundant parentheses around simple identifiers and literals
-    if (
-        expression &&
-        (expression.type === "Identifier" ||
-            expression.type === "Literal" ||
-            expression.type === "CurrentArgsExpression" ||
-            expression.type === "TemplateLiteral" ||
-            expression.type === "UnaryExpression")
-    ) {
-        // Exception: new (Foo) vs new Foo? No, GML doesn't have `new` operator syntax quirks like that usually.
-        // Exception: (1).toString()? GML doesn't have method calls on literals like JS.
-        // For UnaryExpression, only dangerous if parent is MemberExpression accessing result
-        if (expression.type === "UnaryExpression" && parent.type === "MemberExpression" && parent.object === node) {
-            return false;
-        }
-
-        return true;
-    }
-
-    // For non-ternary cases, only process synthetic parentheses
-    if (!isSynthetic) {
-        if (parent.type === "BinaryExpression" && expression?.type === "BinaryExpression") {
-            const parentInfo = getBinaryOperatorInfo(parent.operator);
-            const childInfo = getBinaryOperatorInfo(expression.operator);
-
-            // If child precedence is strictly higher, parens are redundant
-            // e.g. (a * b) + c -> * > +
-            if (childInfo && parentInfo && childInfo.prec > parentInfo.prec) {
-                // Aggressively strip non-synthetic parentheses for arithmetic operations.
-                if (childInfo.type === "arithmetic") {
-                    return !hasImmediateExplicitArithmeticGrouping(expression);
-                }
-
-                // For comparison operations inside logical expressions, check for consistent grouping style.
-                // If only one operand is parenthesized (e.g. `(a > b) && c`), strip it as noise.
-                // If both operands are parenthesized (e.g. `(a > b) || (c < d)`), preserve the intent.
-                if (
-                    childInfo.type === "comparison" &&
-                    parentInfo.type === "logical" &&
-                    expression === node.expression // verifying we are checking the content
-                ) {
-                    // Check if sibling is parenthesized
-                    const otherOperand = parent.left === node ? parent.right : parent.left;
-                    // We check the raw node in AST to see if it's ParenthesizedExpression
-                    // But print.ts receives the path... wait.
-                    // The `node` variable in `shouldOmitSyntheticParens` is the ParenthesizedExpression itself.
-                    // `parent` is the LogicalExpression.
-
-                    // If parent.left === node, sibling is parent.right.
-                    // But we need to use path-based access to be safe?
-                    // Or just raw node access since we have `parent`.
-                    // Prettier ensures AST nodes are stable.
-
-                    if (otherOperand.type !== "ParenthesizedExpression" || otherOperand.synthetic === true) {
-                        return true;
-                    }
-                }
-            }
-
-            if (shouldFlattenSyntheticBinary(parent, expression, path)) {
-                return true;
-            }
-        }
-
-        return shouldFlattenMultiplicationChain(parent, expression, path);
-    }
-
-    if (parent.type === "CallExpression") {
-        return shouldFlattenSyntheticCall(parent, expression, path);
-    }
-
-    if (parent.type !== "BinaryExpression") {
-        return false;
-    }
-
-    // Same-precedence binary chains (e.g. a + b + c, a && b && c) and
-    // comparisons inside logical tests (e.g. a >= 1 or b < 70) are always
-    // flattened regardless of the _flattenSyntheticNumericParens flag.
-    if (expression?.type === "BinaryExpression" && shouldFlattenSyntheticBinary(parent, expression, path)) {
-        return true;
-    }
-
-    const parentInfo = getBinaryOperatorInfo(parent.operator);
-    if (expression?.type === "BinaryExpression" && parentInfo !== undefined) {
-        const childInfo = getBinaryOperatorInfo(expression.operator);
-
-        if (
-            childInfo !== undefined &&
-            childInfo.prec > parentInfo.prec &&
-            shouldFlattenComparisonLogicalTest(parent, expression, path)
-        ) {
-            return true;
-        }
-
-        if (
-            childInfo !== undefined &&
-            childInfo.type === "arithmetic" &&
-            parentInfo.type === "arithmetic" &&
-            childInfo.prec > parentInfo.prec &&
-            !hasImmediateExplicitArithmeticGrouping(expression)
-        ) {
-            return true;
-        }
-    }
-
-    // Numeric parenthesization (e.g. a + (b * c)) requires explicit opt-in
-    if (!isSyntheticParenFlatteningEnabled(path)) {
-        return false;
-    }
-
-    if (expression?.type === "BinaryExpression" && parentInfo !== undefined) {
-        const childInfo = getBinaryOperatorInfo(expression.operator);
-
-        if (childInfo !== undefined && childInfo.prec > parentInfo.prec) {
-            const numericDecision = evaluateNumericBinaryFlattening(parent, expression, path);
-            if (numericDecision === "allow") {
-                return true;
-            }
-            if (numericDecision === "deny") {
-                return false;
-            }
-        }
-
-        if (shouldFlattenMultiplicationChain(parent, expression, path)) {
-            return true;
-        }
-    }
-
-    if (parent.operator !== "+") {
-        return false;
-    }
-
-    if (!binaryExpressionContainsString(parent)) {
-        return false;
-    }
-
-    let depth = 1;
-    while (true) {
-        const ancestor = safeGetParentNode(path, depth - 1);
-        if (!ancestor) {
-            return false;
-        }
-
-        if (ancestor.type === "ParenthesizedExpression" && ancestor.synthetic !== true) {
-            return true;
-        }
-
-        depth += 1;
-    }
-}
-
-// For ternary expressions, omit unnecessary parentheses around simple identifiers
-// or member expressions in the guard/test position. This mirrors the previous
-// inline logic that only trimmed parentheses when they added no semantic value,
-// keeping the formatter's promise of minimal grouping while avoiding precedence
-// changes in more complex logical expressions.
-function shouldFlattenTernaryTest(parentKey, expression) {
-    if (parentKey !== "test") {
-        return false;
-    }
-
-    const expressionType = expression?.type;
-    if (!expressionType) {
-        return false;
-    }
-
-    return (
-        expressionType === "Identifier" ||
-        expressionType === "MemberDotExpression" ||
-        expressionType === "MemberIndexExpression"
-    );
-}
-
-function shouldWrapTernaryExpression(path) {
-    const node = safeGetPathValue(path);
-    if (node && node.__skipTernaryParens) {
-        return false;
-    }
-
-    // Do not wrap ternary expressions in parentheses by default.
-    // The golden fixture tests expect ternary expressions to remain unwrapped
-    // in variable declarations, assignments, and template strings.
-    return false;
-}
-
-function printTernaryExpressionNode(_node, path, _options, print) {
-    const testDoc = path.call(print, "test");
-    const consequentDoc = path.call(print, "consequent");
-    const alternateDoc = path.call(print, "alternate");
-
-    const ternaryDoc = group([testDoc, indent([line, "? ", consequentDoc, line, ": ", alternateDoc])]);
-
-    return shouldWrapTernaryExpression(path) ? concat(["(", ternaryDoc, ")"]) : ternaryDoc;
-}
-
-function hasImmediateExplicitArithmeticGrouping(node) {
-    if (!node || node.type !== "BinaryExpression") {
-        return false;
-    }
-
-    for (const operand of [node.left, node.right]) {
-        if (operand?.type !== "ParenthesizedExpression" || operand.synthetic === true) {
-            continue;
-        }
-
-        const innerExpression = operand.expression;
-        if (
-            (node.operator === "*" || node.operator === "/") &&
-            innerExpression?.type === "BinaryExpression" &&
-            (innerExpression.operator === "*" || innerExpression.operator === "/")
-        ) {
-            continue;
-        }
-
-        return true;
-    }
-
-    return false;
-}
-
-function shouldStripStandaloneAdditiveParentheses(parent, parentKey, expression) {
-    if (!parent || !expression) {
-        return false;
-    }
-
-    if (!isNumericComputationNode(expression)) {
-        return false;
-    }
-
-    const isBinaryExpression = expression.type === "BinaryExpression";
-    if (isBinaryExpression && binaryExpressionContainsString(expression)) {
-        return false;
-    }
-
-    const operatorText = isBinaryExpression ? Core.getNormalizedOperator(expression) : null;
-    const isMultiplicativeExpression =
-        isBinaryExpression && operatorText !== null && MULTIPLICATIVE_BINARY_OPERATORS.has(operatorText);
-
-    switch (parent.type) {
-        case "VariableDeclarator": {
-            return parentKey === "init";
-        }
-        case "AssignmentExpression": {
-            return parentKey === "right";
-        }
-        case "ExpressionStatement": {
-            return parentKey === "expression";
-        }
-        case "ReturnStatement":
-        case "ThrowStatement": {
-            return parentKey === "argument";
-        }
-        case "BinaryExpression": {
-            if (isMultiplicativeExpression) {
-                return false;
-            }
-            if (parent.operator === "+") {
-                return parentKey === "left" || parentKey === "right";
-            }
-
-            if (parent.operator === "-") {
-                return parentKey === "left";
-            }
-
-            return false;
-        }
-        default: {
-            return false;
-        }
-    }
-}
-
-// Synthetic parenthesis flattening only treats select call expressions as
-// numeric so we avoid unwrapping macro invocations that expand to complex
-// expressions. The list is intentionally small and can be extended as other
-// numeric helpers require the same treatment.
-
-function binaryExpressionContainsString(node) {
-    if (!node || node.type !== "BinaryExpression") {
-        return false;
-    }
-
-    if (node.operator !== "+") {
-        return false;
-    }
-
-    return expressionIsStringLike(node.left) || expressionIsStringLike(node.right);
-}
-
-function unwrapParensForInitializer(node) {
-    let current = node;
-    while (current?.type === "ParenthesizedExpression") {
-        current = current.expression;
-    }
-    return current;
-}
-
-function shouldBreakVariableInitializerOnAssignmentLine(node): boolean {
-    if (!node || node.type !== "VariableDeclarator") {
-        return false;
-    }
-
-    const initializer = unwrapParensForInitializer(node.init);
-    return initializer?.type === "BinaryExpression" && binaryExpressionContainsString(initializer);
-}
-
-function unwrapParenthesizedExpression(childPath, print) {
-    const childNode = childPath.getValue();
-    if (childNode?.type === "ParenthesizedExpression") {
-        return childPath.call((innerPath) => unwrapParenthesizedExpression(innerPath, print), "expression");
-    }
-
-    return print();
-}
-
-function printSimpleDeclaration(leftDoc, rightDoc) {
-    return rightDoc ? [leftDoc, " = ", rightDoc] : leftDoc;
-}
-
-// prints empty parens with dangling comments
-function printEmptyParens(path, options) {
-    return group(
-        [
-            "(",
-            indent([printDanglingCommentsAsGroup(path, options, (comment) => !comment.attachToBrace)]),
-            ifBreak(line, "", { groupId: Symbol.for("emptyparen") }),
-            ")"
-        ],
-        { id: Symbol.for("emptyparen") }
-    );
-}
-
-// prints an empty block with dangling comments
-function printEmptyBlock(path, options) {
-    const node = path.getValue();
-    const inlineCommentDoc = maybePrintInlineEmptyBlockComment(path, options);
-
-    if (inlineCommentDoc) {
-        return inlineCommentDoc;
-    }
-
-    const comments = Core.getCommentArray(node);
-    const hasPrintableComments = comments.some(Core.isCommentNode as any);
-
-    if (hasPrintableComments) {
-        const sourceMetadata = resolvePrinterSourceMetadata(options);
-        const shouldAddTrailingBlankLine =
-            sourceMetadata.originalText !== null &&
-            hasBlankLineBetweenLastCommentAndClosingBrace(node, sourceMetadata, sourceMetadata.originalText);
-
-        const trailingDocs = [hardline, "}"];
-        if (shouldAddTrailingBlankLine) {
-            trailingDocs.unshift(lineSuffixBoundary as any, hardline as any);
-        }
-
-        const inlineDangling = printDanglingComments(path, options, (comment) => comment.attachToBrace);
-        const groupedDangling = printDanglingCommentsAsGroup(path, options, (comment) => !comment.attachToBrace);
-        if (groupedDangling) {
-            return ["{", inlineDangling, indent([groupedDangling]), ...trailingDocs];
-        }
-
-        // an empty block with comments
-        return ["{", inlineDangling, ...trailingDocs];
-    } else {
-        return "{}";
-    }
-}
-
-function maybePrintInlineEmptyBlockComment(path, options) {
-    const node = path.getValue();
-    if (!node) {
-        return null;
-    }
-
-    const comments = Core.getCommentArray(node);
-    if (comments.length === 0) {
-        return null;
-    }
-
-    const inlineIndex = findInlineBlockCommentIndex(comments);
-
-    if (inlineIndex < 0) {
-        return null;
-    }
-
-    const comment = comments[inlineIndex];
-    const commentLeadingWS =
-        typeof comment === "object" && comment !== null && "leadingWS" in comment
-            ? (comment as { leadingWS: unknown }).leadingWS
-            : undefined;
-    const commentTrailingWS =
-        typeof comment === "object" && comment !== null && "trailingWS" in comment
-            ? (comment as { trailingWS: unknown }).trailingWS
-            : undefined;
-    const leadingSpacing = getInlineBlockCommentSpacing(commentLeadingWS, " ");
-    const trailingSpacing = getInlineBlockCommentSpacing(commentTrailingWS, " ");
-
-    return [
-        "{",
-        leadingSpacing,
-        path.call((commentPath) => printComment(commentPath, options), "comments", inlineIndex),
-        trailingSpacing,
-        "}"
-    ];
-}
-
-function findInlineBlockCommentIndex(comments) {
-    let inlineIndex = -1;
-
-    for (const [index, comment] of comments.entries()) {
-        if (!Core.isCommentNode(comment)) {
-            continue;
-        }
-
-        if (!isInlineEmptyBlockComment(comment)) {
-            return -1;
-        }
-
-        if (inlineIndex !== -1) {
-            return -1;
-        }
-
-        inlineIndex = index;
-    }
-
-    return inlineIndex;
-}
-
-function getInlineBlockCommentSpacing(text, fallback) {
-    if (typeof text !== STRING_TYPE || text.length === 0) {
-        return fallback;
-    }
-
-    return hasLineBreak(text) ? fallback : text;
-}
-
-function shouldFlattenSyntheticBinary(parent, expression, _path) {
-    const parentInfo = getBinaryOperatorInfo(parent.operator);
-    const expressionInfo = getBinaryOperatorInfo(expression.operator);
-
-    if (!parentInfo || !expressionInfo) {
-        return false;
-    }
-
-    const parentKey = safeGetPathName(_path);
-
-    if (parent.operator === expression.operator) {
-        if ((parent.operator === "-" || parent.operator === "/") && parentKey === "right") {
-            return false;
-        }
-        return true;
-    }
-
-    const parentIsAdditive = parent.operator === "+" || parent.operator === "-";
-    const expressionIsAdditive = expression.operator === "+" || expression.operator === "-";
-    if (!parentIsAdditive || !expressionIsAdditive) {
-        return false;
-    }
-
-    // Flatten additive synthetic parentheses when associativity is preserved.
-    // Safe: (a + b) - c, (a - b) + c, a + (b - c), a + (b + c)
-    // Unsafe: a - (b + c), a - (b - c)
-    if (parentKey === "left") {
-        return true;
-    }
-
-    return parentKey === "right" && parent.operator === "+";
-}
-
-function shouldFlattenMultiplicationChain(parent, expression, _path) {
-    const parentInfo = getBinaryOperatorInfo(parent.operator);
-    const expressionInfo = getBinaryOperatorInfo(expression.operator);
-
-    if (!parentInfo || !expressionInfo) {
-        return false;
-    }
-
-    const parentOperandKey = safeGetPathName(_path);
-
-    if (parent.operator === "/" && parentOperandKey === "right") {
-        return false;
-    }
-
-    // Multiplication associativity
-    return (
-        (parent.operator === "*" || parent.operator === "/") &&
-        (expression.operator === "*" || expression.operator === "/")
-    );
-}
-
-function shouldFlattenSyntheticCall(_parent, _expression, _path) {
-    return false;
-}
-
-function shouldFlattenComparisonLogicalTest(parent, expression, _path) {
-    const parentInfo = getBinaryOperatorInfo(parent.operator);
-    const expressionInfo = getBinaryOperatorInfo(expression.operator);
-
-    if (!parentInfo || !expressionInfo) {
-        return false;
-    }
-
-    // Flatten logic inside logic (e.g. `(a && b) || c`) if precedence allows
-    if (parentInfo.type === "logical" && (expressionInfo.type === "comparison" || expressionInfo.type === "logical")) {
-        return true;
-    }
-
-    // Flatten arithmetic inside comparison (e.g. `a < (b * c)`) if precedence allows
-    if (parentInfo.type === "comparison" && expressionInfo.type === "arithmetic") {
-        return true;
-    }
-
-    return false;
-}
-
-function evaluateNumericBinaryFlattening(parent, expression, _path) {
-    const parentInfo = getBinaryOperatorInfo(parent.operator);
-    const expressionInfo = getBinaryOperatorInfo(expression.operator);
-
-    if (!parentInfo || !expressionInfo) {
-        return;
-    }
-
-    // Always flatten standard arithmetic chains if safe (e.g. `a + b * c` where precedence allows)
-    // The caller ensures childInfo.prec > parentInfo.prec before checking "allow"
-    if (parentInfo.type === "arithmetic" && expressionInfo.type === "arithmetic") {
-        // Exception: modulo? No, precedence handles it.
-        return "allow";
-    }
-
-    // Flatten bitwise inside comparison/arithmetic if safe
-    if (parentInfo.type === "bitwise" || expressionInfo.type === "bitwise") {
-        return "allow";
-    }
-}
